# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: UpWatch API
  version: 0.1.0
  description: |
    API de monitoramento do UpWatch.

    Toda funcionalidade da interface passa por aqui — não existe caminho
    privilegiado para a interface gráfica nem API paralela para automação.
    O que a tela faz, um script faz igual.

    ## Autenticação

    Duas formas, aceitas nas mesmas rotas:

    - **Cookie de sessão**, obtido em `POST /api/v1/auth/login`. É o que a
      interface usa.
    - **Token Bearer**, criado em `POST /api/v1/auth/tokens`. É o caminho
      para scripts e integrações.

    O segredo de um token aparece **apenas** na resposta que o cria. Depois
    disso só o prefixo é exposto, e nada permite reconstruí-lo.

    ## Retenção

    Batidas cruas vivem poucos dias; o histórico longo é servido por
    `/rollups`, que responde por meses sem custo proporcional. Uma janela
    de vários meses deve ser consultada em `/rollups`, não em
    `/heartbeats`.
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
    description: A própria instalação
...
paths:
...
  /api/v1/public:
    get:
      tags: [public]
      security: []
      summary: Página pública padrão
      description: |
        Responde a página marcada como padrão, sem slug no caminho.

        Existe porque numa instalação com uma página só o slug repete o
        que o caminho já diz&#58; "/status/estado" é redundante, e
        "/status" é o endereço que se cola em contrato e em rodapé de
        e-mail.

        Devolve 404 quando nenhuma página foi marcada como padrão — a
        instalação não elege uma sozinha.
      responses:
        "200":
          description: Estado da página
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicView" }
        "304": { description: Sem alteração desde a versão do cliente }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/public/feed.atom:
    get:
      tags: [public]
      security: []
      summary: Feed Atom da página padrão
      responses:
        "200":
          description: Documento Atom
          content:
            application/atom+xml:
              schema: { type: string }
        "304": { description: Sem alteração desde a versão do cliente }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/public/{slug}:
    get:
      tags: [public]
      security: []
      summary: Página pública de estado
      description: |
        Sem credencial, de propósito — é o link que se compartilha durante
        uma queda.

        Página inexistente e página desligada devolvem o mesmo 404:
        distinguir as duas confirmaria a existência da página a quem só
        chutou o endereço.

        A resposta traz `Cache-Control` e `ETag`. Revalidar com
        `If-None-Match` devolve 304 — é a única rota do UpWatch desenhada
        para receber tráfego de verdade.
      parameters:
        - $ref: "#/components/parameters/PathSlug"
      responses:
        "200":
          description: Estado da página
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicView" }
        "304": { description: Sem alteração desde a versão do cliente }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/public/{slug}/feed.atom:
    get:
      tags: [public]
      security: []
      summary: Feed Atom dos relatos
      description: |
        Permite acompanhar a página sem cadastrar e-mail, e integrar num
        canal de chat sem que o UpWatch precise saber falar com ele.

        O feed é derivado da mesma visão da página, então não tem como
        mostrar nada que ela não mostre.
      parameters:
        - $ref: "#/components/parameters/PathSlug"
      responses:
        "200":
          description: Documento Atom
          content:
            application/atom+xml:
              schema: { type: string }
        "304": { description: Sem alteração desde a versão do cliente }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: upwatch_session
    bearerToken:
      type: http
      scheme: bearer
  parameters:
...
    PathSlug:
      name: slug
      in: path
      required: true
      schema: { type: string, pattern: "^[a-z0-9]+(-[a-z0-9]+)*$", maxLength: 64 }
      description: Identificador da página na URL
...
  responses:
...
    NotFound:
      description: Registro não encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - unauthorized
                - not_found
                - conflict
                - too_many_requests
                - internal_error
            message: { type: string }
            field:
              type: string
              description: Campo reprovado, quando o erro for de validação
...
    PublicView:
      type: object
      description: |
        A página como sai para quem não tem credencial. Note o que não
        existe aqui&#58; endereço de alvo, causa detectada, identificador
        de monitor e contagem de alvos da instalação.
      properties:
        slug: { type: string }
        title: { type: string }
        description: { type: string }
        time_zone: { type: string }
        status:
          type: string
          enum: [up, down, degraded, unknown]
          description: Pior estado entre os componentes
        impact:
          type: string
          enum: [none, minor, major, critical]
          description: Maior impacto entre os relatos abertos
        window_days: { type: integer, example: 90 }
        generated_at: { type: string, format: date-time }
        groups:
          type: array
          items: { $ref: "#/components/schemas/PublicGroup" }
        announcements:
          type: array
          items: { $ref: "#/components/schemas/PublicAnnouncement" }
    PublicGroup:
      type: object
      properties:
        name:
          type: string
          description: Vazio é o grupo implícito dos componentes sem agrupamento
        monitors:
          type: array
          items: { $ref: "#/components/schemas/PublicMonitor" }
    PublicMonitor:
      type: object
      properties:
        name: { type: string }
        status: { type: string, enum: [up, down, degraded, unknown] }
        uptime_percent:
          type: number
          nullable: true
          description: |
            Ausente quando nada foi observado na janela. Zero afirmaria
            queda total, que é a leitura mais alarmante possível.
        latency_p95_ms:
          type: integer
          format: int64
          nullable: true
          description: Só aparece se a página liberar latência
        history:
          type: array
          items: { $ref: "#/components/schemas/PublicDay" }
    PublicDay:
      type: object
      properties:
        date:
          type: string
          format: date
          description: |
            Dia de calendário, não instante&#58; um carimbo faria cada
            navegador reinterpretá-lo no próprio fuso e a mesma barra
            mudaria de lugar dependendo de quem olha.
        status: { type: string, enum: [up, down, degraded, unknown] }
        uptime_percent: { type: number, nullable: true }
    PublicAnnouncement:
      type: object
      properties:
        title: { type: string }
        impact: { type: string, enum: [none, minor, major, critical] }
        phase:
          type: string
          enum: [investigating, identified, monitoring, resolved]
        components:
          type: array
          items: { type: string }
          description: Rótulos públicos, nunca identificadores
        started_at: { type: string, format: date-time }
        resolved_at: { type: string, format: date-time, nullable: true }
        updates:
          type: array
          items: { $ref: "#/components/schemas/PublicAnnouncementUpdate" }
    PublicAnnouncementUpdate:
      type: object
      properties:
        phase:
          type: string
          enum: [investigating, identified, monitoring, resolved]
        body: { type: string }
        published_at: { type: string, format: date-time }
security:
  - sessionCookie: []
  - bearerToken: []